Crate modules

Runnable modules containing each bundling multiple services and providing a unified configuration

Modules

api

GraphQL based interface to extract metadata from the grid

collector

Metadata aggregator which builds and persists objects for each session

constants

Constant values

gangway

HTTP traffic ingress for all non-internal interactions with the grid

manager

Schedules newly created sessions, assigning them to provisioners

node

Manages a WebDriver instance and translates requests

options

Various options usable by modules

orchestrator

Module to provision new instances of the node module